PP-5321: old capture process clean up
* remove old capture process class (and tests related to it)
* remove old backgroud task
* renamed `CardCaptureMessageProcess` to `CardCaptureProcess`
* removed unused methods in `ChargeDao`
* removed unused config values
* updated README.md (remove env variables, update description of capture process)
