+++ src/main/java/uk/gov/pay/connector/paymentprocessor/service/CardCaptureProcess.java
@@ -1,143 +1,105 @@
 package uk.gov.pay.connector.paymentprocessor.service;
 
-import com.codahale.metrics.MetricRegistry;
-import com.google.common.base.Stopwatch;
-import io.dropwizard.setup.Environment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.slf4j.MDC;
-import uk.gov.pay.connector.app.CaptureProcessConfig;
-import uk.gov.pay.connector.app.ConnectorConfiguration;
-import uk.gov.pay.connector.charge.dao.ChargeDao;
-import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
-import uk.gov.pay.connector.common.exception.ConflictRuntimeException;
+import uk.gov.pay.connector.charge.ChargesAwaitingCaptureMetricEmitter;
+import uk.gov.pay.connector.charge.service.ChargeService;
+import uk.gov.pay.connector.common.exception.IllegalStateRuntimeException;
 import uk.gov.pay.connector.gateway.CaptureResponse;
-import uk.gov.pay.connector.util.RandomIdGenerator;
+import uk.gov.pay.connector.queue.CaptureQueue;
+import uk.gov.pay.connector.queue.ChargeCaptureMessage;
+import uk.gov.pay.connector.queue.QueueException;
 
 import javax.inject.Inject;
 import java.util.List;
-import java.util.Queue;
-import java.util.concurrent.TimeUnit;
-
-import static java.lang.String.format;
 
 public class CardCaptureProcess {
 
-    private static final String HEADER_REQUEST_ID = "X-Request-Id";
-
-    private final Logger logger = LoggerFactory.getLogger(getClass());
-    private final ChargeDao chargeDao;
-    private final CardCaptureService captureService;
-    private final MetricRegistry metricRegistry;
-    private final CaptureProcessConfig captureConfig;
-    private Queue<ChargeEntity> captureQueue;
-    private volatile int totalChargesCaptured;
-    private volatile String runCaptureId;
-    private volatile int readyCaptureQueueSize;
-    private volatile int waitingCaptureQueueSize;
-    private volatile int chargesToCaptureSize;
+    private static final Logger LOGGER = LoggerFactory.getLogger(CardCaptureProcess.class);
+    private final CaptureQueue captureQueue;
+    private final ChargeService chargeService;
+    private CardCaptureService cardCaptureService;
 
     @Inject
-    public CardCaptureProcess(Environment environment, ChargeDao chargeDao, CardCaptureService cardCaptureService,
-                              ConnectorConfiguration connectorConfiguration, Queue<ChargeEntity> captureQueue) {
-        this.chargeDao = chargeDao;
-        this.captureService = cardCaptureService;
-        this.captureConfig = connectorConfiguration.getCaptureProcessConfig();
+    public CardCaptureProcess(CaptureQueue captureQueue,
+                              CardCaptureService cardCaptureService,
+                              ChargeService chargeService,
+                              ChargesAwaitingCaptureMetricEmitter chargesAwaitingCaptureMetricEmitter) {
         this.captureQueue = captureQueue;
-        metricRegistry = environment.metrics();
-        metricRegistry.gauge("gateway-operations.capture-process.queue-size.ready_capture_queue_size", () -> () -> readyCaptureQueueSize);
-        metricRegistry.gauge("gateway-operations.capture-process.queue-size.waiting_capture_queue_size", () -> () -> waitingCaptureQueueSize);
+        this.cardCaptureService = cardCaptureService;
+        this.chargeService = chargeService;
+
+        chargesAwaitingCaptureMetricEmitter.register();
     }
 
-    public void runCapture(int threadNumber) {
-        MDC.put(HEADER_REQUEST_ID, format("runCapture-%s, thread %d", runCaptureId, threadNumber));
+    public void handleCaptureMessages() throws QueueException {
+        List<ChargeCaptureMessage> captureMessages = captureQueue.retrieveChargesForCapture();
+        for (ChargeCaptureMessage message : captureMessages) {
+            try {
+                LOGGER.info("Charge capture message received - [externalChargeId={}] [queueMessageId={}] [queueMessageReceiptHandle={}]",
+                        message.getChargeId(),
+                        message.getQueueMessageId(),
+                        message.getQueueMessageReceiptHandle()
+                );
 
-        Stopwatch responseTimeStopwatch = Stopwatch.createStarted();
-        int captured = 0, skipped = 0, error = 0, failedCapture = 0, total = 0;
+                runCapture(message);
+            } catch (Exception e) {
+                LOGGER.warn("Error capturing charge from SQS message [externalChargeId={}] [queueMessageId={}] [errorMessage={}]",
+                        message.getChargeId(),
+                        message.getQueueMessageId(),
+                        e.getMessage()
+                );
+            }
+        }
+    }
+
+    private void runCapture(ChargeCaptureMessage captureMessage) throws QueueException {
+        String externalChargeId = captureMessage.getChargeId();
+
         try {
-            ChargeEntity charge;
-            while ((charge = getChargeToCapture()) != null) {
-                total = incrementTotalChargesCaptured();
-                if (shouldRetry(charge)) {
-                    try {
-                        logger.info("Capturing [{} of {}] [chargeId={}]", total, chargesToCaptureSize, charge.getExternalId());
-                        CaptureResponse gatewayResponse = captureService.doCapture(charge.getExternalId());
-                        if (gatewayResponse.isSuccessful()) {
-                            captured++;
-                        } else {
-                            logger.info("Failed to capture [chargeId={}] due to: {}", charge.getExternalId(),
-                                    gatewayResponse.getError().get().getMessage());
-                            failedCapture++;
-                        }
-                    } catch (ConflictRuntimeException e) {
-                        logger.info("Another process has already attempted to capture [chargeId={}]. Skipping.", charge.getExternalId());
-                        skipped++;
-                    } catch (Exception e) {
-                        logger.info("Exception [{}] when capturing charge [chargeId={}]. Skipping.", e.getMessage(), charge.getExternalId());
-                        skipped++;
-                    }
-                } else {
-                    captureService.markChargeAsCaptureError(charge.getExternalId());
-                    error++;
-                }
-            }
-        } catch (Exception e) {
-            logger.error("Exception [{}] when running capture at charge [{} of {}]", e.getMessage(), total, chargesToCaptureSize, e);
-        } finally {
-            responseTimeStopwatch.stop();
-            metricRegistry.histogram("gateway-operations.capture-process.running_time").update(responseTimeStopwatch.elapsed(TimeUnit.MILLISECONDS));
-            if (readyCaptureQueueSize > 0) {
-                logger.info("Capture complete [captured={}] [skipped={}] [capture_error={}] [failed_capture={}] [total={}]",
-                        captured, skipped, error, failedCapture, readyCaptureQueueSize);
-            }
-        }
-        MDC.remove(HEADER_REQUEST_ID);
-    }
+            CaptureResponse gatewayResponse = cardCaptureService.doCapture(externalChargeId);
 
-    private synchronized int incrementTotalChargesCaptured() {
-        return ++totalChargesCaptured;
-    }
-
-    public synchronized void loadCaptureQueue() {
-        if (captureQueue.isEmpty()) {
-            runCaptureId = RandomIdGenerator.newId();
-            MDC.put(HEADER_REQUEST_ID, format("runCapture-%s", runCaptureId));
-
-            waitingCaptureQueueSize = chargeDao.countChargesAwaitingCaptureRetry(captureConfig.getRetryFailuresEveryAsJavaDuration());
-
-            List<ChargeEntity> chargesToCapture = chargeDao.findChargesForCapture(captureConfig.getBatchSize(),
-                    captureConfig.getRetryFailuresEveryAsJavaDuration());
-
-            chargesToCaptureSize = chargesToCapture.size();
-
-            if (chargesToCaptureSize < captureConfig.getBatchSize()) {
-                readyCaptureQueueSize = chargesToCaptureSize;
+            if (gatewayResponse.isSuccessful()) {
+                captureQueue.markMessageAsProcessed(captureMessage);
             } else {
-                readyCaptureQueueSize = chargeDao.countChargesForImmediateCapture(captureConfig.getRetryFailuresEveryAsJavaDuration());
+                LOGGER.info(
+                        "Failed to capture [externalChargeId={}] due to: {}",
+                        externalChargeId,
+                        gatewayResponse.getErrorMessage()
+                );
+                handleCaptureRetry(captureMessage);
             }
-
-            if (chargesToCaptureSize > 0) {
-                logger.info("Capturing: {} of {} charges", chargesToCaptureSize, (waitingCaptureQueueSize + readyCaptureQueueSize));
-            }
-
-            captureQueue.addAll(chargesToCapture);
-            MDC.remove(HEADER_REQUEST_ID);
+        } catch (IllegalStateRuntimeException e) {
+            handleCapturedInvalidTransition(captureMessage, e);
         }
     }
 
-    private synchronized ChargeEntity getChargeToCapture() {
-        return captureQueue.poll();
+    private void handleCaptureRetry(ChargeCaptureMessage captureMessage) throws QueueException {
+        boolean shouldRetry = chargeService.isChargeRetriable(captureMessage.getChargeId());
+
+        if (shouldRetry) {
+            LOGGER.info("Charge capture message [{}] scheduled for retry.", captureMessage.getChargeId());
+            captureQueue.scheduleMessageForRetry(captureMessage);
+        } else {
+            cardCaptureService.markChargeAsCaptureError(captureMessage.getChargeId());
+            captureQueue.markMessageAsProcessed(captureMessage);
+        }
     }
 
-    private boolean shouldRetry(ChargeEntity charge) {
-        return chargeDao.countCaptureRetriesForCharge(charge.getId()) < captureConfig.getMaximumRetries();
-    }
+    private void handleCapturedInvalidTransition(ChargeCaptureMessage captureMessage, IllegalStateRuntimeException e) throws QueueException {
+        if (chargeService.isChargeCaptureSuccess(captureMessage.getChargeId())) {
+            LOGGER.info(
+                    "Charge capture message [{}] already captured - marking as processed. [chargeId={}]",
+                    captureMessage.getQueueMessageId(),
+                    captureMessage.getChargeId());
+            captureQueue.markMessageAsProcessed(captureMessage);
+            return;
+        }
 
-    int getReadyCaptureQueueSize() {
-        return readyCaptureQueueSize;
-    }
-
-    int getWaitingCaptureQueueSize() {
-        return waitingCaptureQueueSize;
+        LOGGER.info(
+                "Capture process non-success illegal state transition for message {} [chargeId={}]",
+                captureMessage.getQueueMessageId(),
+                captureMessage.getChargeId());
+        throw e;
     }
 }
