+++ src/main/resources/config/config.yaml
@@ -92,12 +92,6 @@ executorServiceConfig:
 
 captureProcessConfig:
   backgroundProcessingEnabled: ${BACKGROUND_PROCESSING_ENABLED:-true}
-  schedulerInitialDelayInSeconds: ${CAPTURE_PROCESS_SCHEDULER_INITIAL_DELAY_SECONDS:-20}
-  schedulerRandomIntervalMinimumInSeconds: ${CAPTURE_PROCESS_SCHEDULER_RANDOM_INTERVAL_MINIMUM_SECONDS:-10}
-  schedulerRandomIntervalMaximumInSeconds: ${CAPTURE_PROCESS_SCHEDULER_RANDOM_INTERVAL_MAXIMUM_SECONDS:-20}
-
-  batchSize: ${CAPTURE_PROCESS_BATCH_SIZE:-10}
-  schedulerThreads: ${SCHEDULER_THREADS:-10}
 
   # The below effectively get multiplied together. In order to handle how
   # certain payment gateways do things, it is extremely desirable to keep these
@@ -106,7 +100,6 @@ captureProcessConfig:
   # e931b4dab25284acedeb6d59d4bfd3d29e454b82 for more details.
   retryFailuresEvery: ${CAPTURE_PROCESS_RETRY_FAILURES_EVERY:-60 minutes}
   maximumRetries: ${CAPTURE_PROCESS_MAXIMUM_RETRIES:-96}
-  captureUsingSQS: ${CAPTURE_USING_SQS_FEATURE_FLAG:-false}
 
   failedCaptureRetryDelayInSeconds: ${CAPTURE_PROCESS_FAILED_CAPTURE_RETRY_DELAY_IN_SECONDS:-3600}
   queueSchedulerThreadDelayInSeconds: ${CAPTURE_PROCESS_QUEUE_SCHEDULER_THREAD_DELAY_IN_SECONDS:-1}
