--- src/test/java/uk/gov/pay/connector/it/gatewayclient/GatewaySocketErrorIT.java
@@ -1,48 +1,0 @@
-package uk.gov.pay.connector.it.gatewayclient;
-
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.pay.connector.it.dao.DatabaseFixtures;
-import uk.gov.pay.connector.paymentprocessor.service.CardCaptureProcess;
-import uk.gov.pay.connector.rules.GuiceAppWithPostgresRule;
-
-import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
-import static com.github.tomakehurst.wiremock.client.WireMock.post;
-import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
-import static io.dropwizard.testing.ConfigOverride.config;
-import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThat;
-import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.CAPTURE_APPROVED_RETRY;
-
-@RunWith(MockitoJUnitRunner.class)
-public class GatewaySocketErrorIT extends BaseGatewayITest {
-
-    @Rule
-    public GuiceAppWithPostgresRule app = new GuiceAppWithPostgresRule(
-            config("smartpay.urls.test", "http://localhost:" + port + "/pal/servlet/soap/Payment"));
-
-    @Before
-    public void setUp() {
-        setupLoggerMockAppender();
-    }
-
-    @Test
-    public void shouldFailCaptureWhenErrorFromGateway() {
-        stubFor(
-                post(urlPathEqualTo("/pal/servlet/soap/Payment"))
-                        .willReturn(aResponse().withStatus(404))
-        );
-
-        DatabaseFixtures.TestCharge testCharge = createTestCharge(app.getDatabaseTestHelper());
-        app.getInstanceFromGuiceContainer(CardCaptureProcess.class).loadCaptureQueue();
-        app.getInstanceFromGuiceContainer(CardCaptureProcess.class).runCapture(1);
-
-        assertLastGatewayClientLoggingEventContains(
-                String.format("Gateway returned unexpected status code: 404, for gateway url=http://localhost:%s/pal/servlet/soap/Payment with type test", port));
-        assertThat(app.getDatabaseTestHelper().getChargeStatus(testCharge.getChargeId()), is(CAPTURE_APPROVED_RETRY.getValue()));
-    }
-}
