--- src/test/java/uk/gov/pay/connector/it/gatewayclient/GatewaySocketReadTimeoutIT.java
@@ -1,41 +1,0 @@
-package uk.gov.pay.connector.it.gatewayclient;
-
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.pay.connector.it.dao.DatabaseFixtures;
-import uk.gov.pay.connector.paymentprocessor.service.CardCaptureProcess;
-import uk.gov.pay.connector.rules.GuiceAppWithPostgresRule;
-
-import static io.dropwizard.testing.ConfigOverride.config;
-import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThat;
-import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.CAPTURE_APPROVED_RETRY;
-
-@RunWith(MockitoJUnitRunner.class)
-public class GatewaySocketReadTimeoutIT extends BaseGatewayITest {
-
-    @Rule
-    public GuiceAppWithPostgresRule app = new GuiceAppWithPostgresRule(
-            config("smartpay.urls.test", "http://localhost:" + port + "/pal/servlet/soap/Payment"),
-            config("customJerseyClient.readTimeout", "5ms"));
-
-    @Before
-    public void setUp() {
-        setupLoggerMockAppender();
-    }
-
-    @Test
-    public void shouldFailCaptureWhenConnectionTimeoutFromGateway() {
-        DatabaseFixtures.TestCharge testCharge = createTestCharge(app.getDatabaseTestHelper());
-        setupGatewayStub().respondWithTimeoutWhenCapture();
-        app.getInstanceFromGuiceContainer(CardCaptureProcess.class).loadCaptureQueue();
-        app.getInstanceFromGuiceContainer(CardCaptureProcess.class).runCapture(1);
-
-        assertLastGatewayClientLoggingEventContains(
-                String.format("Connection timed out error for gateway url=http://localhost:%s/pal/servlet/soap/Payment", port));
-        assertThat(app.getDatabaseTestHelper().getChargeStatus(testCharge.getChargeId()), is(CAPTURE_APPROVED_RETRY.getValue()));
-    }
-}
