+++ src/test/java/uk/gov/pay/connector/it/resources/CardResourceCaptureWithSqsQueueIT.java
@@ -32,10 +32,7 @@ import static uk.gov.pay.connector.common.model.api.ExternalChargeState.EXTERNAL
 
 @RunWith(DropwizardJUnitRunner.class)
 @DropwizardConfig(app = ConnectorApp.class, config = "config/test-it-config.yaml",
-        configOverrides = {
-                @ConfigOverride(key = "captureProcessConfig.captureUsingSQS", value = "true"),
-                @ConfigOverride(key = "captureProcessConfig.backgroundProcessingEnabled", value = "false")
-        },
+        configOverrides = {@ConfigOverride(key = "captureProcessConfig.backgroundProcessingEnabled", value = "false")},
         withDockerSQS = true
 )
 public class CardResourceCaptureWithSqsQueueIT extends ChargingITestBase {
